[fix] Change return value of key_cmp to int instead of int64_t

## Proposed changes

This pull request introduces the Tiresia test, which tests keys and values of all possible sizes using BerkeleyDB as the source of truth to test Parallax.  
This PR closes issues #34, #35, and #37.


## Types of changes

What types of changes does your code introduce?
_Put an `x` in the boxes that apply_

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update (if none of the other choices apply)

## Checklist

_Put an `x` in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code._

- [x] I have read the [CONTRIBUTING](https://carvgit.ics.forth.gr/gxanth/parallax/blob/master/CONTRIBUTING.md) doc
- [x] Lint and unit tests pass locally with my changes
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have added necessary documentation (if appropriate)

See merge request storage/parallax!68

+++ lib/btree/btree.h
@@ -498,7 +498,7 @@ void find_key(struct lookup_operation *get_op);
 int8_t delete_key(db_handle *handle, void *key, uint32_t size);
 
 void init_key_cmp(struct key_compare *key_cmp, void *key_buf, char key_format);
-int64_t key_cmp(struct key_compare *key1, struct key_compare *key2);
+int key_cmp(struct key_compare *key1, struct key_compare *key2);
 int prefix_compare(char *l, char *r, size_t prefix_size);
 
 void recover_L0(struct db_descriptor *db_desc);
